pyproject.toml
[project]
name = "siteofshift"
version = "0.1.0"
description = "Provider site-of-care optimization project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib",
  "numpy",
  "pandas",
  "pyyaml",
  "pytest",
  "scikit-learn",
  "streamlit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/siteofshift"]

[tool.hatch.build.targets.sdist]
include = [
  "src/siteofshift",
  "config",
  "tests",
  "README.md",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]